apt: Revert 8d27f68 partially
It turns out that the issue was that the Pin-Priority for testing
etc. was too high, 800 is too high, so is 900, just setting it to 200
and 100/50 for unstable/experimental works. So do that.

Presumably the details of how pinning happens internally and the
assigned weights changed in Jessie.

When I ran "dug" before 8d27f68 it had a *huge* conflict resolution
calculation that never ended, this works just fine, and I can see
e.g. bitcoind from unstable.