squid: Add a config that in theory works for replacing webproxy.corp.booking.com in bash
I'm adding this as a start of having a config where I can not need to
do:

    https_proxy= git pull

When I'm not in the office. I was trying to do this automagically with
squid but that doesn't seem to work (see comment in the file itself).

Now I just need to do this in /etc/squid3/squid.conf.d:

    sudo ln -sf elsewhere.conf location.conf
    sudo ln -sf office.conf location.conf

And:

    sudo /etc/init.d/squid3 reload

And things should Just Work, in theory.