Add the squid3 autoconfigure script
This is the second half of what I started in 7d96c31, this seems to just
work! I'm now switching back & forth between my work & home config and
"curl" at al just work now with my
/home/avar/g/avar-dotfiles-work/.bashrc.d/11-booking-proxy.sh pointing
to localhost:3128

The stuff I'm echo-ing ends up in:

    tail -f /var/log/daemon.log | grep nm-dispatcher

I'm also changing the network config in GNOME for myself in this
script, so now the last stuff I had un-automated when I connect to
this is automated!