Also run the if-up.d script in if-post-down.d & always tweak dconf
I'm putting it into if-post-down.d because I could e.g. disconnect from
work eth0 and go to wlan0 on the guest network or whatever. This covers
all the basis.

Also always tweak the dconf settings just in case I've been fiddling
with them or whatever, shouldn't strictly speaking be needed, but ...